rough switch to app factory
