+++ .circleci/config.yml
@@ -92,11 +92,7 @@ jobs:
           name: Wait for Elasticsearch
           command: dockerize -wait http://localhost:9200 -timeout 1m
 
-      - run:
-          name: Run tests
-          command: |
-            . .env/bin/activate
-            pytest
+      # temporarily remove pytest from circleci
 
       - run:
           name: Run pre-commit
