+++ bin/run.sh
@@ -1,2 +1,2 @@
 python cli.py cf_startup
-gunicorn --access-logfile - --error-logfile - --log-level info --timeout 300 -k gevent -w 9 webservices.rest:app
+gunicorn --access-logfile - --error-logfile - --log-level info --timeout 300 -k gevent -w 9 'webservices.rest:create_app()'
