+++ tests/integration/test_ao_elasticsearch.py
@@ -1,6 +1,6 @@
 from tests.common import ElasticSearchBaseTest
 from webservices.resources.legal import UniversalSearch, REQUESTOR_TYPES
-from webservices.rest import api
+from webservices.api_setup import api
 from datetime import datetime
 from webservices.legal_docs import TEST_SEARCH_ALIAS
