+++ tests/integration/test_cases_elasticsearch.py
@@ -2,7 +2,7 @@ from datetime import datetime
 from webservices.resources.legal import ALL_DOCUMENT_TYPES
 from tests.common import ElasticSearchBaseTest, ALL_INDICES, document_dictionary
 from webservices.legal_docs import TEST_SEARCH_ALIAS
-from webservices.rest import api
+from webservices.api_setup import api
 from webservices.resources.legal import UniversalSearch
 
 import unittest.mock as mock
