+++ tests/test_aggregates.py
@@ -5,7 +5,8 @@ from tests.common import ApiBaseTest, assert_dicts_subset
 from webservices.utils import get_current_cycle
 
 from webservices import schemas
-from webservices.rest import db, api
+from webservices.common.models import db
+from webservices.api_setup import api
 from webservices.resources.aggregates import (
     CommunicationCostByCandidateView,
     ElectioneeringByCandidateView,
