+++ tests/test_committees.py
@@ -6,8 +6,8 @@ from tests import factories
 from tests.common import ApiBaseTest
 
 from webservices import utils
-from webservices.rest import db
-from webservices.rest import api
+from webservices.common.models import db
+from webservices.api_setup import api
 from webservices.resources.committees import CommitteeList
 from webservices.resources.committees import CommitteeView
 from webservices.resources.committees import CommitteeHistoryProfileView
