+++ tests/test_filters.py
@@ -3,7 +3,7 @@ from tests import factories
 from tests.common import ApiBaseTest
 
 from webservices import filters
-from webservices.rest import api
+from webservices.api_setup import api
 from webservices.common import models
 from webservices.resources.dates import CalendarDatesView
 from webservices.resources.committees import CommitteeList
