+++ tests/test_inaugural_donations.py
@@ -1,7 +1,7 @@
 from tests import factories
 from tests.common import ApiBaseTest, assert_dicts_subset
 
-from webservices.rest import api
+from webservices.api_setup import api
 from webservices.resources.totals import InauguralDonationsView
