+++ tests/test_large_aggregates.py
@@ -1,7 +1,7 @@
 from tests import factories
 from tests.common import ApiBaseTest
 
-from webservices.rest import api
+from webservices.api_setup import api
 from webservices.resources.large_aggregates import EntityReceiptDisbursementTotalsView
