+++ tests/test_national_party.py
@@ -3,7 +3,7 @@ import sqlalchemy as sa
 
 from tests import factories
 from tests.common import ApiBaseTest
-from webservices.rest import api
+from webservices.api_setup import api
 from webservices.schemas import NationalPartyTotalsSchema
 from webservices.resources.national_party import (
     NationalParty_ScheduleAView,
