+++ tests/test_presidential.py
@@ -2,7 +2,7 @@ import datetime
 from tests import factories
 from tests.common import ApiBaseTest
 
-from webservices.rest import api
+from webservices.api_setup import api
 from webservices.resources.presidential import (
     PresidentialByCandidateView,
     PresidentialByStateView,
