+++ tests/test_reports.py
@@ -7,8 +7,8 @@ from tests import factories
 from tests.common import ApiBaseTest
 
 from webservices import schemas
-from webservices.rest import db
-from webservices.rest import api
+from webservices.common.models import db
+from webservices.api_setup import api
 from webservices.resources.reports import (
     ReportsView,
     CommitteeReportsView,
