+++ tests/test_sched_c.py
@@ -4,7 +4,7 @@ import datetime
 from tests import factories
 from tests.common import ApiBaseTest
 
-from webservices.rest import api
+from webservices.api_setup import api
 from webservices.schemas import ScheduleCSchema
 from webservices.resources.sched_c import ScheduleCView, ScheduleCViewBySubId
