+++ tests/test_sched_f.py
@@ -1,7 +1,7 @@
 import datetime
 from tests import factories
 from tests.common import ApiBaseTest
-from webservices.rest import api
+from webservices.api_setup import api
 from webservices.schemas import ScheduleFSchema
 from webservices.resources.sched_f import ScheduleFView, ScheduleFViewBySubId
