+++ tests/test_utils.py
@@ -8,11 +8,11 @@ from tests import factories
 from tests.common import ApiBaseTest
 
 from webservices import args
-from webservices import rest
+from flask import current_app
 from webservices import sorting
 from webservices.resources import candidate_aggregates
 from webservices.resources import elections
-from webservices.rest import db
+from webservices.common.models import db
 from webservices import utils
 from webservices.common import models
 
@@ -308,9 +308,10 @@ class TestSort(ApiBaseTest):
 
 class TestArgs(TestCase):
     def test_currency(self):
-        with rest.app.test_request_context('?dollars=$24.50'):
-            parsed = flaskparser.parser.parse({'dollars': args.Currency()}, request, location='query')
-            self.assertEqual(parsed, {'dollars': 24.50})
+        if current_app.config['TESTING']:
+            with current_app.test_request_context('?dollars=$24.50'):
+                parsed = flaskparser.parser.parse({'dollars': args.Currency()}, request, location='query')
+                self.assertEqual(parsed, {'dollars': 24.50})
 
 
 class TestEnvVarSplit(TestCase):
