+++ webservices/legal_docs/advisory_opinions.py
@@ -1,7 +1,7 @@
 from collections import defaultdict
 import logging
 import re
-from webservices.rest import db
+from webservices.common.models import db
 from webservices.utils import (
     create_es_client,
     DateTimeEncoder,
